--- a/.clang-tidy
+++ b/.clang-tidy
@@ -82,6 +82,7 @@ Checks: >
     -readability-make-member-function-const,
     -readability-redundant-string-init,
     -readability-non-const-parameter,
+    -readability-container-contains,
     -readability-static-accessed-through-instance
     
 WarningsAsErrors: '*'