--- a/.github/workflows/osrm-backend.yml
+++ b/.github/workflows/osrm-backend.yml
@@ -270,7 +270,7 @@ jobs:
           - name: gcc-11-release
             continue-on-error: false
             node: 20
-            runs-on: ubuntu-24.04
+            runs-on: ubuntu-22.04
             BUILD_TOOLS: ON
             BUILD_TYPE: Release
             CCOMPILER: gcc-11
@@ -420,13 +420,24 @@ jobs:
         # Linux dev packages
         if [ "${ENABLE_CONAN}" != "ON" ]; then
           sudo apt-get update -y
-          sudo apt-get install -y libbz2-dev libxml2-dev libzip-dev liblua5.2-dev libboost-all-dev
+          sudo apt-get install -y libbz2-dev libxml2-dev libzip-dev liblua5.2-dev
           if [[ "${CCOMPILER}" != clang-* ]]; then
             sudo apt-get install -y ${CXXCOMPILER}
           fi
           if [[ "${ENABLE_COVERAGE}" == "ON" ]]; then
             sudo apt-get install -y lcov
           fi
+          
+          # Install Boost 1.85
+          BOOST_VERSION="1.85.0"
+          BOOST_VERSION_UNDERSCORE="${BOOST_VERSION//./_}"
+          wget -q https://boostorg.jfrog.io/artifactory/main/release/${BOOST_VERSION}/source/boost_${BOOST_VERSION_UNDERSCORE}.tar.gz
+          tar xzf boost_${BOOST_VERSION_UNDERSCORE}.tar.gz
+          cd boost_${BOOST_VERSION_UNDERSCORE}
+          ./bootstrap.sh
+          ./b2 install
+          cd ..
+          rm -rf boost_${BOOST_VERSION_UNDERSCORE}*
         fi
 
         # TBB