--- a/.github/workflows/osrm-backend.yml
+++ b/.github/workflows/osrm-backend.yml
@@ -434,8 +434,8 @@ jobs:
           wget -q https://boostorg.jfrog.io/artifactory/main/release/${BOOST_VERSION}/source/boost_${BOOST_VERSION_UNDERSCORE}.tar.gz
           tar xzf boost_${BOOST_VERSION_UNDERSCORE}.tar.gz
           cd boost_${BOOST_VERSION_UNDERSCORE}
-          ./bootstrap.sh
-          ./b2 install
+          sudo ./bootstrap.sh
+          sudo ./b2 install
           cd ..
           rm -rf boost_${BOOST_VERSION_UNDERSCORE}*
         fi