--- a/.github/workflows/osrm-backend.yml
+++ b/.github/workflows/osrm-backend.yml
@@ -277,27 +277,27 @@ jobs:
           #   CXXCOMPILER: g++-11
           #   ENABLE_BENCHMARKS: ON
 
-          - name: conan-linux-release-node
-            build_node_package: true
-            continue-on-error: false
-            node: 20
-            runs-on: ubuntu-22.04
-            BUILD_TYPE: Release
-            CCOMPILER: clang-13
-            CXXCOMPILER: clang++-13
-            ENABLE_CONAN: ON
-            NODE_PACKAGE_TESTS_ONLY: ON
+          # - name: conan-linux-release-node
+          #   build_node_package: true
+          #   continue-on-error: false
+          #   node: 20
+          #   runs-on: ubuntu-22.04
+          #   BUILD_TYPE: Release
+          #   CCOMPILER: clang-13
+          #   CXXCOMPILER: clang++-13
+          #   ENABLE_CONAN: ON
+          #   NODE_PACKAGE_TESTS_ONLY: ON
 
-          - name: conan-linux-debug-node
-            build_node_package: true
-            continue-on-error: false
-            node: 20
-            runs-on: ubuntu-22.04
-            BUILD_TYPE: Debug
-            CCOMPILER: clang-13
-            CXXCOMPILER: clang++-13
-            ENABLE_CONAN: ON
-            NODE_PACKAGE_TESTS_ONLY: ON
+          # - name: conan-linux-debug-node
+          #   build_node_package: true
+          #   continue-on-error: false
+          #   node: 20
+          #   runs-on: ubuntu-22.04
+          #   BUILD_TYPE: Debug
+          #   CCOMPILER: clang-13
+          #   CXXCOMPILER: clang++-13
+          #   ENABLE_CONAN: ON
+          #   NODE_PACKAGE_TESTS_ONLY: ON
 
           # - name: conan-macos-x64-release-node
           #   build_node_package: true