--- a/cmake/warnings.cmake
+++ b/cmake/warnings.cmake
@@ -79,3 +79,5 @@ add_warning(tautological-compare)
 add_warning(trampolines)
 # these warnings are not enabled by default
 # no_warning(name-of-warning)
+no_warning(deprecated-comma-subscript)
+no_warning(comma-subscript)