--- a/include/util/static_assert.hpp
+++ b/include/util/static_assert.hpp
@@ -14,8 +14,8 @@ template <typename It, typename Value> inline void static_assert_iter_value()
 
 template <typename It, typename Category> inline void static_assert_iter_category()
 {
-    // using IterCategoryType = typename std::iterator_traits<It>::iterator_category;
-    // static_assert(std::is_base_of<Category, IterCategoryType>::value, "");
+    using IterCategoryType = typename std::iterator_traits<It>::iterator_category;
+    static_assert(std::is_base_of<Category, IterCategoryType>::value, "");
 }
 
 } // namespace osrm::util